Move User related methods to WifiDevice superclass
