+++ src/main/java/com/oktawski/iotserver/devices/relay/Relay.java
@@ -1,8 +1,6 @@
 package com.oktawski.iotserver.devices.relay;
 
-import com.fasterxml.jackson.annotation.JsonBackReference;
 import com.oktawski.iotserver.superclasses.WifiDevice;
-import com.oktawski.iotserver.user.models.User;
 
 import javax.persistence.*;
 
@@ -10,18 +8,6 @@ import javax.persistence.*;
 @Table(name = "relays")
 public class Relay extends WifiDevice {
 
-    @JsonBackReference
-    @ManyToOne(fetch = FetchType.LAZY, optional = false)
-    @JoinColumn(name = "user_id")
-    private User user = null;
-
     public Relay(){}
 
-    public User getUser() {
-        return user;
-    }
-
-    public void setUser(User user) {
-        this.user = user;
-    }
 }
