+++ src/main/java/com/oktawski/iotserver/devices/thermometer/Thermometer.java
@@ -12,22 +12,9 @@ public class Thermometer extends WifiDevice {
 
     private double temperature;
 
-    @JsonBackReference
-    @ManyToOne(fetch = FetchType.LAZY, optional = false)
-    @JoinColumn(name = "user_id")
-    private User user = null;
-
     public Thermometer(){}
 
     public double getTemperature() {
         return temperature;
-    }
-
-    public User getUser() {
-        return user;
-    }
-
-    public void setUser(User user) {
-        this.user = user;
     }
 }
